requirements.txt
langchain
langchain-ollama
langchain-community
psycopg2-binary
python-dotenv
tabulate
colorama
pymongo
customtkinter
pyinstaller
